multicast resume

+++ src/info5/sar/multicast/LamportClock.java
@@ -1,17 +1,17 @@
 package info5.sar.multicast;
 
 public class LamportClock {
-    private long time = 0;
+    private int time = 0;
 
-    public synchronized long tick() {
+    public synchronized int tick() {
         return ++time;
     }
 
-    public synchronized void update(long other) {
+    public synchronized void update(int other) {
         time = Math.max(time, other) + 1;
     }
 
-    public synchronized long get() {
+    public synchronized int get() {
         return time;
     }
 }
